Epubsearch, version 1.0.0
Description: This script searches for a string within all the EPUB
ebooks in a directory tree. The titles and file names of the EPUBs that
match, and their matching paragraphs, will be output. The search string
is case-insensitive and ignores the width of spacing between words. The
search string can be a regexp.